Refactor: flatten module layout and introduce runtimes + bench scaffolding
